+++ docs/astro.config.mjs
@@ -1,11 +1,12 @@
 import { defineConfig } from "astro/config";
 import preact from "@astrojs/preact";
 import react from "@astrojs/react";
+import sitemap from "@astrojs/sitemap";
 
 // https://astro.build/config
 export default defineConfig({
-  integrations: [preact(), react()],
-  site: `https://openapi-typescript.pages.dev`,
+  integrations: [preact(), react(), sitemap()],
+  site: `https://openapi-ts.pages.dev`,
   vite: {
     define: {
       "import.meta.env.VITE_ALGOLIA_APP_ID": JSON.stringify(process.env.ALGOLIA_APP_ID ?? ""),